+++ CHANGELOG.md
@@ -5,6 +5,7 @@ PATCH
 
 ### Payments
 * [Added] Added API bindings for Naver Pay.
+* [Added] Added API bindings for PAYCO.
 
 ## 26.8.0 2026-08-24
 ### Payments